# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "servers": [
        {
            "url": "https://app.agm-sa.net/api"
        }
    ],
    "paths": {
...
        "/provider/employees/{id}/suggestCourses": {
            "put": {
                "tags": [
                    "Provider Employees"
                ],
                "summary": "Suggest Courses for employee ",
                "operationId": "7744e96c2cb2e1017a8b9a9ccb0f861b",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "course_id": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "response": [],
                                            "success": true
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}